deep_gaze/requirements.txt
opencv-python==3.4.0.14
opencv-contrib-python==3.4.2.17
tensorflow_gpu==1.4.0
matplotlib==3.0.2
scikit-learn==0.19.1
cividis
mahotas==1.4.7
pandas==0.24.1
seaborn==0.9.0
Cython==0.29.3
pysaliency
dilli==0.2.8.2
